backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.3.0
phonenumbers==8.13.26
APScheduler==3.11.2
sendgrid==6.12.5
